rename package that I forgot to rename, doh
