+++ ps1rfid/config.go
@@ -20,7 +20,7 @@
  *
  */
 
-package cfg
+package ps1rfid
 
 import (
 	"fmt"
